Fancy Server Procedures
Use traits and a few special types to make server procedures a little
bit easier to write, and harder to screw up. The `ServerProcedure` trait
returns a type with a lifetime annotation, making it easier for the
compiler to tell if you are about to leak memory.

This was a problem I noticed with the plain `door_return` function -- I
could easily pass it a vec.as_bytes(), and the client would get the
right data, but my vec would never get dropped. So I would end up with a
memory leak that rustc was unable to catch because door_return doesn't
behave like a normal function.