Separate door creation / runtime concerns
The ServerProcedure trait can be implemented for a type, and this will
endow it with some methods that can create a door server (represented by
a RawFd).

Later, other code can use this door server object to advertise the door
on the filesystem, revoke it, or (eventually) read and write its runtime
parameters.